login functionality added with a user profile page
